fix: 修复 PR 审查问题 - predict_tool 注册、数据/依赖/模型一致性、补 API 与测试
1. predict_tool 注册到 Tool 注册中心（核心）：agent/tools/__init__.py 导入真实
   predict_tool 并加入 _REAL_REGISTRY，ACTIVE_REGISTRY 改为优先真实实现，接通 Agent 预测能力。
2. requirements.txt：prophet 改为可选注释依赖（Windows 编译风险），代码已有
   PROPHET_AVAILABLE + sklearn 降级方案。
3. anomaly_scenarios.py：发电机定子过热注入上限 125→145℃，超过 threshold_high(140) 以触发告警。
4. 删除误提交的 commit_msg.txt。
5. 新增 backend/api/predict.py（GET /api/v1/predict、/predict/model）并在 main.py 注册路由，
   供前端/C 调用。
7. models.py TimeseriesData.id 改 BigInteger；init.sql 加复合主键 (id, recorded_at)
   满足 TimescaleDB hypertable 唯一约束需含分区列的要求。
9. generate_data.py create_alarm_records 按 score>0.5 分 high/medium，与 ThresholdDetector 一致。
10. 新增 tests/algorithm/test_prediction.py（metrics/降级/trend 判定）
    与 test_model_selector.py（高/低负荷工况划分）。

另：合并 develop 同步 A 的 P0-5 基础设施（Tool 注册中心、agent/graph、rag、frontend 等），
冲突统一取 develop 版本（A 已集成 B 的 P0 修复）；risk_scorer 0.3/0.6 阈值已正确，
agent_service._extract_diagnosis 阈值由 A 侧同步。

+++ backend/database/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Double, DateTime, ForeignKey, Text, ARRAY
+from sqlalchemy import Column, Integer, BigInteger, String, Double, DateTime, ForeignKey, Text, ARRAY
 from sqlalchemy.sql import func
 from backend.database.connection import Base
 
@@ -32,7 +32,8 @@ class SensorPoint(Base):
 class TimeseriesData(Base):
     __tablename__ = "timeseries_data"
 
-    id = Column(Integer, primary_key=True, autoincrement=True)
+    # 与 init.sql 的 BIGSERIAL 一致，使用 64 位整型，避免 32 位 Integer 溢出
+    id = Column(BigInteger, primary_key=True, autoincrement=True)
     device_id = Column(Integer, ForeignKey("device.id"))
     sensor_id = Column(Integer, ForeignKey("sensor_point.id"))
     value = Column(Double, nullable=False)
